+++ CHANGES.md
@@ -4,6 +4,7 @@
 - The Full Day view no longer shows days that are hidden (the Today button is also hidden to prevent jumping to a hidden day).
 - Renamed the CSS class :current-year-selector to :year-selected.
 - Added support to shown the years that contain events in the years down-down menu (new CSS class :year-has-events).
+- Added support to search by Group and Url in the Search dialog.
 - 
 
 ### Version 1.1.1: